extract goals from commander
